Resolve issue on Windows where `npm --prefix` does not work as expected.
The `npm --prefix` switch is being ignored when developers run the command on Windows. This seems to be a known issue: https://github.com/npm/cli/issues/4511

The workaround is to create a nested shell with the correct working directory which should work well for developers on Mac and Windows.