Refer to 'First Name' and 'Last Name' consistently.
There are some inconsistencies in how names are referred to (first name, last name, given name, surname, etc).

UIs are referring to names as 'First Name' and 'Last Name' and it has been agreed that the source code will reflect the same convention. This change updates the source code accordingly.

The only exception is around the user attributes that are provided by Microsoft when the authentication extensions are called. Microsoft are referring to the user's name as 'Given Name' and 'Surname'.

At this point database columns will not be renamed.